fix duplications
